Reboot the website as digital garden.
After 6 years since my last update, this website will be turned in a
digital garden[1].

I hope this commit is the beginning of my journey as a digital gardener.

As a way of making more commitment more public, this message is also
documenting the beginning of this new version of my personal website.

I hope new content will be added soon.

[1] Gardens evolve and are always a work in progress